fix menu-btn hover style, change props name in navbar, refactor popup
