+++ src/components/PopupWithForm/PopupWithForm.js
@@ -1,5 +1,4 @@
-import React from 'react';
-import PopupContainer from '../PopupContainer/PopupContainer';
+import React, { useEffect, useRef } from 'react';
 import { ReactComponent as CloseIcon } from '../../images/close.svg';
 import './PopupWithForm.css';
 
@@ -8,22 +7,46 @@ const PopupWithForm = (props) => {
     title,
     linkName,
     children,
-    onClose,
+    closePopup,
     // onSubmit,
     changePopup,
   } = props;
 
+  const popupContainer = useRef(null);
+
+  useEffect(() => {
+    document.addEventListener('click', handleClick, false);
+    document.addEventListener('keydown', handleEsc, false);
+
+    return () => {
+      document.removeEventListener('click', handleClick, false);
+      document.removeEventListener('keydown', handleEsc, false);
+    };
+  }, []);
+
+  const handleClick = (evt) => {
+    if (
+      popupContainer.current
+      && !popupContainer.current.contains(evt.target)
+    ) {
+      closePopup();
+    }
+  };
+
+  const handleEsc = (evt) => {
+    if (evt.key && evt.key === 'Escape') {
+      closePopup();
+    }
+  };
+
   return (
     <div className="popup popup_opened">
-      <PopupContainer
-        className="popup__container"
-        closeHandler={onClose}
-      >
+      <div className="popup__container" ref={popupContainer}>
         <button
           type="button"
           className="popup__close-btn"
           title="Закрыть"
-          onClick={onClose}
+          onClick={closePopup}
         >
           <CloseIcon />
         </button>
@@ -43,7 +66,7 @@ const PopupWithForm = (props) => {
             {linkName}
           </button>
         </p>
-      </PopupContainer>
+      </div>
     </div>
   );
 }
